Add typed deserialization for Message
abstract classes annotated with JsonSerializable will create functions
to encode and decode a concrete type of Message by putting a type key in
the `Command` field. The name is derived from the prefix of the concrete
class. All inherited classes must be named <custom><parent>, like
SomeMessage. Key will be "Some" for abstract class "Message".

Message offers fromTypedJson and toTypedJson that will add the "Command"
field. This field is not accessible from the created instances and only
exists in the json.